Update matching.py
I wanted to inform you that I made some changes to your code to solve some problems I encountered when using it.

First, I changed the parts of the code that used np.float to np.float64 and installed numpy 1.24. This step was necessary to resolve a conflict that was causing problems with versions of numpy prior to 1.20.

In addition, I had to import the bbox_ious function into the code. This correction was necessary because, without it, the code would fail.

I hope that these changes will improve the efficiency and functionality of the code. If you have any questions or concerns regarding these changes, please do not hesitate to let me know.